feat: add offline mode (#81)
* feat: add offline mode

* fix: determining script directory when paths are relative

* test: for offline mode feature

* docs: add offline mode feature documentation

* chore(build): auto-generate vimdoc

---------

+++ scripts/neovim_install.sh
@@ -5,7 +5,6 @@ set -eo pipefail
 
 # Create a temporary directory to handle any remote nvim data location things
 temp_dir=$(mktemp -d 2>/dev/null || mktemp -d -t 'neovim_download')
-cd "$temp_dir" || exit 1
 
 cleanup_function() {
 	# Function to delete the directory, change "/path/to/directory" to the actual directory path
@@ -24,6 +23,7 @@ Options:
   -d       Specify directory for storing Neovim binaries.
            NOTE: Installation would happen in 'nvim-downloads' subdirectory.
   -f       Force installation. Would overwrite any existing installation.
+  -o       Offline mode. Assume release is already downloaded.
   -h       Display this help message and exit.
 EOM
 }
@@ -49,37 +49,42 @@ function download() {
 }
 
 # Install on Linux using AppImage
-function download_decompress_neovim_linux_appimage() {
-	local nvim_version="$1"
-	local download_url="https://github.com/neovim/neovim/releases/download/${nvim_version}/nvim.appimage"
+function setup_neovim_linux_appimage() {
+	local nvim_release_name="nvim-$1-linux.appimage"
+	local nvim_appimage_temp_path="$temp_dir/$nvim_release_name"
 
-	echo "Downloading Neovim for Linux (AppImage)..."
-	download "$download_url" "$temp_dir/nvim.appimage"
+	if [ ! -e "$nvim_version_dir/$nvim_release_name" ]; then
+		echo "Expected release to be present at $nvim_version_dir/$nvim_release_name. Aborting..."
+		exit 1
+	fi
+
+	cp "$nvim_version_dir/$nvim_release_name" "$nvim_appimage_temp_path"
 
 	echo "Extracting Neovim binary..."
-	chmod u+x "$temp_dir/nvim.appimage"
-	"$temp_dir/nvim.appimage" --appimage-extract >/dev/null
+	chmod u+x "$nvim_appimage_temp_path"
+	"$nvim_appimage_temp_path" --appimage-extract >/dev/null
 
 	echo "Finishing up installing Neovim..."
-	rm -rf "$nvim_version_dir"
 	mkdir -p "$nvim_version_dir"/bin
 	mv -f "$temp_dir/squashfs-root"/* "$nvim_version_dir"
 	ln -sf "$nvim_version_dir"/AppRun "$nvim_binary"
 }
 
 # Function to download and decompress Neovim binary for macOS
-function download_decompress_neovim_macOS() {
-	local nvim_version="$1"
-	local download_url="https://github.com/neovim/neovim/releases/download/${nvim_version}/nvim-macos.tar.gz"
+function setup_neovim_macos() {
+	local nvim_release_name="nvim-$1-macos.tar.gz"
+	local nvim_macos_tar_path="$temp_dir/$nvim_release_name"
+	cp "$nvim_version_dir/$nvim_release_name" "$nvim_macos_tar_path"
 
-	echo "Downloading Neovim for macOS..."
-	download "$download_url" "$temp_dir/nvim-macos.tar.gz"
+	if [ ! -e "$nvim_version_dir/$nvim_release_name" ]; then
+		echo "Expected release to be present at $nvim_version_dir/$nvim_release_name"
+		exit 1
+	fi
 
 	echo "Extracting Neovim binary..."
-	tar -xzf "$temp_dir/nvim-macos.tar.gz" -C "$temp_dir"
+	tar -xzf "$nvim_macos_tar_path" -C "$temp_dir"
 
 	echo "Finishing up Neovim installation..."
-	rm -rf "$nvim_version_dir"
 	mkdir -p "$nvim_version_dir"
 	mv -f "$temp_dir"/nvim-macos/* "$nvim_version_dir"
 
@@ -88,11 +93,6 @@ function download_decompress_neovim_macOS() {
 
 # Function to install Neovim
 function install_neovim() {
-	# Check if Neovim is available globally in the system's $PATH
-	# if ! $force_installation; then
-	# 	check_neovim_in_path
-	# fi
-
 	# Check if the specified download directory exists
 	if [[ ! -d $remote_nvim_dir ]]; then
 		echo "Remote neovim directory does not exist. Creating it now..."
@@ -112,33 +112,42 @@ function install_neovim() {
 		fi
 
 		mkdir -p "$nvim_version_dir"
+		local os
+		os=$(uname)
 
-		# Check if either curl or wget is available on the system
-		if command -v curl &>/dev/null; then
-			downloader="curl"
-		elif command -v wget &>/dev/null; then
-			downloader="wget"
+		if [ "$offline_mode" == true ]; then
+			echo "Operating in offline mode. Will not download Neovim release"
 		else
-			echo "Error: This script requires either curl or wget to be installed."
-			exit 1
+			# Check if either curl or wget is available on the system
+			if command -v curl &>/dev/null; then
+				downloader="curl"
+			elif command -v wget &>/dev/null; then
+				downloader="wget"
+			else
+				echo "Error: This script requires either curl or wget to be installed."
+				exit 1
+			fi
+
+			"$download_neovim_script" -o "$os" -v "$nvim_version" -d "$nvim_version_dir"
 		fi
 
 		# Install Neovim based on the detected OS
-		if [[ "$(uname)" == "Linux" ]]; then
-			download_decompress_neovim_linux_appimage "$nvim_version"
-		elif [[ "$(uname)" == "Darwin" ]]; then
-			download_decompress_neovim_macOS "$nvim_version"
+		if [[ $os == "Linux" ]]; then
+			setup_neovim_linux_appimage "$nvim_version"
+		elif [[ $os == "Darwin" ]]; then
+			setup_neovim_macos "$nvim_version"
 		else
 			echo "Unsupported operating system: $(uname)"
 			exit 1
 		fi
+
 	fi
 
 	echo "Neovim $nvim_version can be accessed at $nvim_binary"
 }
 
 # Parse command-line options
-while getopts "v:d:h:f" opt; do
+while getopts "v:d:h:fo" opt; do
 	case $opt in
 	v)
 		nvim_version="$OPTARG"
@@ -148,6 +157,9 @@ while getopts "v:d:h:f" opt; do
 		;;
 	f)
 		force_installation=true
+		;;
+	o)
+		offline_mode=true
 		;;
 	h)
 		display_help
@@ -171,4 +183,7 @@ if [[ -z $nvim_version || -z $remote_nvim_dir ]]; then
 	exit 1
 fi
 
+SCRIPT_DIR=$(cd -- "$(dirname -- "${BASH_SOURCE[0]}")" &>/dev/null && pwd)
+download_neovim_script="$SCRIPT_DIR/neovim_download.sh"
+cd "$temp_dir" || exit 1
 install_neovim
